Make BPS exit with code of child process
BPS uses subprocesses to execute the commands responsible for generating
and updating the quantum graph.  If such a  command fails, though the
actual exit code is shown in the error message the *BPS command* itself
always exits with the error code 1.  This prevents WMS from taking
appropriate actions, e.g., retry the job running the BPS command with a
larger memory if OOM happened.  Made changes so the BPS command exit
code is the same as the exit code of its child processes if any of them
fail for some reason.